added a script to check whether hSC agrees with our stuff

+++ modules/cl_interpolator.py
@@ -17,7 +17,8 @@ class ClInterpolator(object):
         """
 
         # Ells below the rannge
-        ls_pre=np.geomspace(2, lb[0],nb_dex_extrap_lo*np.log10(lb[0]/2.))
+        # B.H. included int casting
+        ls_pre=np.geomspace(2, lb[0],int(nb_dex_extrap_lo*np.log10(lb[0]/2.)))
         # Ells in range
         ls_mid=(lb[:-1,None]+(np.arange(nrb)[None,:]*np.diff(lb)[:,None]/nrb)).flatten()[1:]
         # Ells above range
